requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.16.0
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==42.0.2
daphne==4.1.2
Django==5.1.7
django-anymail==12.0
django-cors-headers==3.14.0
django-environ==0.12.0
django-phonenumber-field==7.3.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
environ==1.0
google-api-core==2.22.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.21.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
packaging==24.1
pexpect==4.9.0
phonenumbers==8.13.30
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-http-client==3.3.7
pytz==2024.1
redis==5.2.1
requests==2.32.3
rsa==4.9
sendgrid==6.11.0
service-identity==24.2.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
stripe==11.4.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
zope.interface==7.2
